refactor(rust): extract models.rs with all shared structs
